+++ src/OpenColorIO/ops/NoOp/NoOps.cpp
@@ -31,7 +31,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include <OpenColorIO/OpenColorIO.h>
 
-#include "ops/Allocation/AllocationOp.h"
+#include "ops/allocation/AllocationOp.h"
 #include "NoOps.h"
 #include "OpBuilders.h"
 #include "Op.h"
