+++ src/OpenColorIO/ops/allocation/AllocationOp.cpp
@@ -30,7 +30,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #include <OpenColorIO/OpenColorIO.h>
 
-#include "ops/Allocation/AllocationOp.h"
+#include "ops/allocation/AllocationOp.h"
 #include "ops/Log/LogOps.h"
 #include "ops/Matrix/MatrixOps.h"
 #include "Op.h"
