+++ src/OpenColorIO/ops/allocation/AllocationOp.h
